Convert a model output into a data frame for Comet/GAVI Analysis

Given a Nimue or Squire simulation object with MCMC fits, this function produces a data frame of Rt over time and the equivalent Beta values, and if an Nimue object, vaccine related parameters. It also extends this data frame for the next 240 days for use in forecasting.

Usage

prepare_input_json_df(out, ox_interventions = NULL)

Arguments

out

Nimue or squire model output, must have MCMC results in pmcmc_results slot.

ox_interventions

Data frame of government interventions, used to add graphical elements, if NULL then element is not added to the data frame.

Value

A tibble containing the necessary data to run Comet or other analyses.

Details

Currently can only take fits to reported deaths and not excess mortality.

Examples

#apply this process to the standard Afghanistan fit
df <- prepare_input_json_df(
   afg_fit,
   ox_interventions = NULL
   )
#> Warning: 'mod_gen(...)' is deprecated; please use 'mod_gen$new(...)' instead
#can be written to a json as needed
if (FALSE) {
   writeLines(
      jsonlite::toJSON(df, pretty = TRUE),
      "input_params.json"
      )
}