Fit a nimue model to the given data using a scan fitting method

Taken from OJs code. Iteratively adjusts the Rt_rw parameters based on maximising likelihood of a short period of time in the data.

Usage

scan_fit(
  res,
  data = NULL,
  width = 2,
  n_span = 8,
  width_end = NULL,
  repeats = 5,
  data_window = 4,
  min_data_window = 3
)

Arguments

res

A nimue simulation object

data

The data to fit to, for standard fits must have date and death columns; for lmic fits must have date, deaths and cases; and for excess mortality it must have deaths, week_start and week_end. Default = NULL, which uses the data in the fit

width

The algorithm will check values up to +- width of the current value

n_span

How many values between the width to check

width_end

The final width to check, so that the algorithm to move to a finer resolution as the proceeds, default NULL means the same as width

repeats

How many times to iterate through the process

data_window

How many data points to include onwards from the date the Rt effect comes into effect, default = NULL.

min_data_window

If an effects data window is less than this value then we skip it. Useful since the scan has issues with effects with only 1 or 2 data points, default = 3.

Value

Model object tuned to the data

Details

Only works with excess fits for now.

Uses furrr for simulations later so can be run in parallel with future::plan(future::multisession(), .cleanup = TRUE)